+++ include/aex/dev/chardevice.hpp
@@ -5,6 +5,7 @@
 #include "aex/errno.hpp"
 #include "aex/mem.hpp"
 #include "aex/optional.hpp"
+#include "aex/types.hpp"
 
 #include <stddef.h>
 #include <stdint.h>
@@ -19,8 +20,8 @@ namespace AEX::Dev {
         virtual error_t open(CharHandle* handle, int mode);
         virtual error_t close(CharHandle* handle);
 
-        virtual optional<uint32_t> read(CharHandle* handle, void* ptr, uint32_t len);
-        virtual optional<uint32_t> write(CharHandle* handle, const void* ptr, uint32_t len);
+        virtual optional<ssize_t> read(CharHandle* handle, void* ptr, size_t len);
+        virtual optional<ssize_t> write(CharHandle* handle, const void* ptr, size_t len);
 
         virtual bool isatty();
     };
