+++ include/aex/dev/charhandle.hpp
@@ -3,6 +3,7 @@
 #include "aex/mem/smartptr.hpp"
 #include "aex/mutex.hpp"
 #include "aex/optional.hpp"
+#include "aex/types.hpp"
 
 #include <stddef.h>
 #include <stdint.h>
@@ -18,8 +19,8 @@ namespace AEX::Dev {
         CharHandle(Mem::SmartPointer<CharDevice> chrhndl);
         ~CharHandle();
 
-        optional<uint32_t> read(void* ptr, uint32_t len);
-        optional<uint32_t> write(const void* ptr, uint32_t len);
+        optional<ssize_t> read(void* ptr, size_t len);
+        optional<ssize_t> write(const void* ptr, size_t len);
 
         bool isatty();
